Use localhost addresses for URLs in Homepage widgets
It seems like using unbound as my DNS resolver broke something with the
SSL certificates and meant that Bert couldn't connect to himself through
the externally-visible name. I think. I don't know why or how that
happened, but using localhost for all of homepage's API calls seems to
fix the problem of big red error boxes on the homepage website.