Remove unnecessary swapDevices on Alex-NixOS
Hopefully this will fix the problem with /swapfile staying in old roots
when using impermanence. I thought the swap file shouldn't be staying in
the first place, but I was confusing a swap partition with /swapfile.
Since I have a dedicated swap partition on Alex, I don't need /swapfile
as well.