Use systemd RestartMode=direct on transmission to stop transmission-ntfy from failing
For some reason, transmission fails a few times before it works, but it
always works in the end. We don't want transmission-ntfy to stop and
give up (because one of its dependencies failed, it won't even restart
itself), so setting RestartMode=direct means that transmission-ntfy
never even knows that transission failed a few times.