Limit tasks for FileFlows server and allow frequent restarts
The FileFlows server might have a memory leak, since it seem to be
constantly creating new threads. After about 1000 threads, the server
locks up and becomes inaccessible through the web UI. This change will
limit the server to 1000 threads and restart it whenever it crashes due
to the thread limit.